Se agrega fragment para materiales. Se modifica el home para que vaya al listado de plantas. Se modifica la lógica del boton + de la pantalla para ir al alta correspondiente
