+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -5,13 +5,13 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_home"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="@string/menu_home" />
-        <item
             android:id="@+id/nav_plant"
-            android:icon="@drawable/ic_menu_gallery"
+            android:icon="@drawable/ic_menu_camera"
             android:title="@string/menu_plant" />
+        <item
+            android:id="@+id/nav_material"
+            android:icon="@drawable/ic_menu_gallery"
+            android:title="@string/menu_material" />
         <item
             android:id="@+id/nav_report"
             android:icon="@drawable/ic_menu_slideshow"
