+++ app/src/main/res/values/strings.xml
@@ -13,8 +13,8 @@
     <string name="menu_header_desc">Navigation header</string>
 
     <!-- Opciones de menu -->
-    <string name="menu_home">Home</string>
     <string name="menu_plant">Plantas</string>
+    <string name="menu_material">Materiales</string>
     <string name="menu_report">Reportes</string>
     <string name="menu_sign_out">Cerrar Sesión</string>
